requirements.txt
jieba
gensim
tqdm


nltk
scikit-image
scipy
torchvision
matplotlib
imageio

graphviz
torchviz